move clientid to sync so syncid can rely on it
