+++ src/embed/dispatch.rs
@@ -1,11 +1,11 @@
 use crate::dag;
 use crate::embed::connection;
 use crate::kv::idbstore::IdbStore;
-use crate::kv::{Store, StoreError};
+use crate::kv::Store;
+use crate::sync;
 use crate::util::rlog;
 use crate::util::rlog::LogContext;
 use crate::util::to_debug;
-use crate::util::uuid::uuid;
 use async_std::sync::{channel, Mutex, Receiver, Sender};
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicU32, Ordering};
@@ -124,7 +124,7 @@ async fn do_open(conns: &mut ConnMap, req: &Request) -> Response {
         Ok(store) => Box::new(store),
     };
 
-    let client_id = init_client_id(kv.as_ref(), req.lc.clone())
+    let client_id = sync::client_id::init(kv.as_ref(), req.lc.clone())
         .await
         .map_err(to_debug)?;
 
@@ -162,56 +162,5 @@ async fn do_debug(conns: &ConnMap, req: &Request) -> Response {
     match req.data.as_string().as_deref() {
         Some("open_dbs") => Ok(JsValue::from_str(&to_debug(conns.keys()))),
         _ => Err("Debug command not defined".into()),
-    }
-}
-
-#[derive(Debug)]
-enum InitClientIdError {
-    CommitErr(StoreError),
-    GetErr(StoreError),
-    InvalidUtf8(std::string::FromUtf8Error),
-    OpenErr(StoreError),
-    PutClientIdErr(StoreError),
-}
-
-async fn init_client_id(s: &dyn Store, lc: LogContext) -> Result<String, InitClientIdError> {
-    use InitClientIdError::*;
-
-    const CID_KEY: &str = "sys/cid";
-    let cid = s.get(CID_KEY).await.map_err(GetErr)?;
-    if let Some(cid) = cid {
-        let s = String::from_utf8(cid).map_err(InvalidUtf8)?;
-        return Ok(s);
-    }
-    let wt = s.write(lc).await.map_err(OpenErr)?;
-    let uuid = uuid();
-    wt.put(CID_KEY, &uuid.as_bytes())
-        .await
-        .map_err(PutClientIdErr)?;
-    wt.commit().await.map_err(CommitErr)?;
-    Ok(uuid)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::kv::memstore::MemStore;
-    use crate::util::rlog::LogContext;
-
-    #[async_std::test]
-    async fn test_init_client_id() {
-        let ms = Box::new(MemStore::new());
-        let cid1 = init_client_id(ms.as_ref(), LogContext::new())
-            .await
-            .unwrap();
-        let cid2 = init_client_id(ms.as_ref(), LogContext::new())
-            .await
-            .unwrap();
-        assert_eq!(cid1, cid2);
-        let ms = Box::new(MemStore::new());
-        let cid3 = init_client_id(ms.as_ref(), LogContext::new())
-            .await
-            .unwrap();
-        assert_ne!(cid1, cid3);
     }
 }
